Add files via upload
This is only a super duper basic version, I'll upload a much better version later on. I was using a file I created in another project to test the writing/reading methods, so I will also add a function to create a userData file later on, amongst many other things. Just wanted to give a little show of faith :D